Fix crash after flinging out the leftmost task card under LTR layout
This bug was introduced in commit "Fix crash when page scroll is not
generated". I have mentioned "possibilities exist that the taskview is
added before generating the page scroll". But unfortunately, it is 100%
that taskviews get removed before removing the page scroll. Fix it by
using min() to avoid any kind of IndexOutOfBoundsException.

Change-Id: Icbdab5ca9fe14a04f858e7475ec2df6a0eca0a8b
Signed-off-by: LibXZR <i@xzr.moe>